requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.0.0
aiosqlite>=0.20.0
bcrypt>=4.0.0
PyJWT>=2.8.0
